services/logging_auditing.py
# your_logging_and_auditing_module.py

import logging
import os

class YourLoggerImplementation:
    def __init__(self):
        # Initialize the logger
        self.logger = logging.getLogger(__name__)
        self.logger.setLevel(logging.DEBUG)

        # Create a file handler and set the level to DEBUG
        self.file_handler = logging.FileHandler('app.log')
        self.file_handler.setLevel(logging.DEBUG)

        # Create a formatter and add it to the file handler
        formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
        self.file_handler.setFormatter(formatter)

        # Add the file handler to the logger
        self.logger.addHandler(self.file_handler)

    def log_message(self, message):
        # Log a message
        self.logger.debug(message)

    def audit_event(self, event):
        # Log an audit event
        self.logger.info(f"AUDIT: {event}")

    def close_logger(self):
        # Close the file handler to release the file
        self.file_handler.close()
