tests/test_logging_auditing.py
# test_logging_and_auditing.py

import pytest
from services.logging_auditing import YourLoggerImplementation
import os


@pytest.fixture(scope='session')
def cleanup_log_file(request):
    # Setup: Delete the log file if it exists before the session
    log_file = 'app.log'
    if os.path.exists(log_file):
        os.remove(log_file)

    # Teardown: Add a finalizer to ensure the log file is deleted after the session
    def finalizer():
        if os.path.exists(log_file):
            os.remove(log_file)

    request.addfinalizer(finalizer)


@pytest.fixture
def logger(request, cleanup_log_file):
    # Create the logger instance
    logger = YourLoggerImplementation()

    # Add a finalizer to ensure the logger is closed after the tests
    def finalizer():
        logger.file_handler.close()

    request.addfinalizer(finalizer)

    return logger

# ... rest of the code remains the same ...


class LoggingAndAuditingTester:
    def __init__(self, logger):
        self.logger = logger

    def function_that_generates_log_entry(self, message):
        self.logger.log_message(message)

    def is_log_entry_present(self, expected_log_message):
        log_entries = self._get_log_entries()
        return any(expected_log_message in entry for entry in log_entries)

    def function_that_generates_log_entry_with_details(self, details):
        self.logger.audit_event(details)

    def is_log_entry_accurate(self, expected_log_details):
        log_entries = self._get_log_entries()
        return any(expected_log_details in entry for entry in log_entries)

    def function_that_triggers_auditing_feature(self, event):
        self.logger.audit_event(event)

    def is_auditing_feature_recorded(self, expected_audit_event):
        log_entries = self._get_log_entries()
        return any(expected_audit_event in entry for entry in log_entries)

    def _get_log_entries(self):
        with open('app.log', 'r') as log_file:
            return log_file.readlines()


def test_log_entry_presence(logger):
    tester = LoggingAndAuditingTester(logger)
    tester.function_that_generates_log_entry("Test Log Message")
    assert tester.is_log_entry_present("Test Log Message")


def test_log_entry_accuracy(logger):
    tester = LoggingAndAuditingTester(logger)
    tester.function_that_generates_log_entry_with_details("Test Log Details")
    assert tester.is_log_entry_accurate("Test Log Details")


def test_audit_feature(logger):
    tester = LoggingAndAuditingTester(logger)
    tester.function_that_triggers_auditing_feature("Test Audit Event")
    assert tester.is_auditing_feature_recorded("Test Audit Event")


if __name__ == "__main__":
    pytest.main()
